profile search fix
fixed the profile search to be in the navbar
